build-ros-images/pyproject.toml
[project]
name = "build-ros-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dagger-io>=0.20.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
]
